% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{alpha_01}
\alias{alpha_01}
\title{Maps from R^n to the open simplex.}
\usage{
alpha_01(alpha.inv)
}
\arguments{
\item{alpha.inv}{a vector in R^n.}
}
\value{
a vector on the (n+1) open simplex.
}
\description{
Maps from R^n to the open simplex.
}
\references{
See \url{https://cran.r-project.org/web/packages/SALTSampler/vignettes/SALTSampler.html}
}
