% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{alpha_real}
\alias{alpha_real}
\title{Maps a (n+1) open simplex to R^n.}
\usage{
alpha_real(alpha)
}
\arguments{
\item{alpha}{vector of weights that live on an open simplex.}
}
\value{
a projection of 'alpha' onto R^n.
}
\description{
Maps a (n+1) open simplex to R^n.
}
