% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.r
\name{dpool}
\alias{dpool}
\title{Obtain the pooled density}
\usage{
dpool(x, D, alpha, log = FALSE)
}
\arguments{
\item{x}{value at which to evaluate the density.}

\item{D}{list containing the density functions.}

\item{alpha}{vector of the same size as the length of 'D' with weights.}

\item{log}{logical. Whether to return the log-density.}
}
\value{
the log-pooled probability density/mass at 'x'.
}
\description{
Obtain the pooled density
}
