% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.r
\name{dpoolnorm}
\alias{dpoolnorm}
\title{Normalised pooled density.}
\usage{
dpoolnorm(x, D, alpha, log = FALSE)
}
\arguments{
\item{x}{numeric. values at which to evaluate the density.}

\item{D}{list containing the density functions.}

\item{alpha}{vector of weights, which must lie in a simplex (can be a scalar).}

\item{log}{logical. Whether to return the log-density.}
}
\value{
numeric. the normalised (log) density.
}
\description{
Normalised pooled density.
}
