% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_elicitator.r
\name{elicit_beta_mean_cv}
\alias{elicit_beta_mean_cv}
\title{Elicit hyperparameters of a Beta distribution from mean and variance or
mean and coefficient of variation}
\usage{
elicit_beta_mean_cv(m0, v0 = NULL, cv = 1)
}
\arguments{
\item{m0}{the desired mean, a scalar between 0 and 1.}

\item{v0}{the desired mean, a scalar between 0 and 1.}

\item{cv}{the desired coefficient of variation, a scalar larger than 0.}
}
\value{
a vector containing the elicited shape parameters.
}
\description{
Elicit hyperparameters of a Beta distribution from mean and variance or
mean and coefficient of variation
}
