% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_elicitator.r
\name{elicit_beta_mean_maxent}
\alias{elicit_beta_mean_maxent}
\title{Elicit hyperparameters of a maximum-entropy Beta distribution with a given mean}
\usage{
elicit_beta_mean_maxent(m0)
}
\arguments{
\item{m0}{target mean}
}
\value{
a vector containing the elicited shape parameters.
}
\description{
Elicit hyperparameters of a maximum-entropy Beta distribution with a given mean
}
