% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_elicitator.r
\name{elicit_beta_median_iq}
\alias{elicit_beta_median_iq}
\title{Elicit hyperparameters of a Beta distribution from median and upper quantile}
\usage{
elicit_beta_median_iq(m, d, q = 0.9)
}
\arguments{
\item{m}{the desired median, a scalar between 0 and 1.}

\item{d}{the distance between \code{m} and the desired upper quantile.}

\item{q}{the desired probability level, a scalar between 0 and 1.}
}
\value{
a vector containing the elicited shape parameters.
}
\description{
Elicit hyperparameters of a Beta distribution from median and upper quantile
}
