% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_elicitator.r
\name{elicit_beta_quantiles}
\alias{elicit_beta_quantiles}
\title{Elicit hyperparameters of a Beta distribution from lower and upper quantiles}
\usage{
elicit_beta_quantiles(l, u, alpha = 0.95)
}
\arguments{
\item{l}{intended lower quantile}

\item{u}{intended upper quantile}

\item{alpha}{the desired probability level, a scalar between 0 and 1.}
}
\value{
a vector containing the elicited shape parameters.
}
\description{
Elicit hyperparameters of a Beta distribution from lower and upper quantiles
}
