% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_and_KL.r
\name{entropy_beta}
\alias{entropy_beta}
\title{Entropy of a Beta distribution}
\usage{
entropy_beta(a, b)
}
\arguments{
\item{a}{Shape parameter. Scalar larger than zero.}

\item{b}{Shape parameter. Scalar larger than zero.}
}
\value{
the entropy, a scalar.
}
\description{
Entropy of a Beta distribution
}
