% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_and_KL.r
\name{entropy_gamma}
\alias{entropy_gamma}
\title{Entropy of a Gamma distribution}
\usage{
entropy_gamma(a, b)
}
\arguments{
\item{a}{Shape parameter. Scalar larger than zero.}

\item{b}{Rate parameter. Scalar larger than zero.}
}
\value{
the entropy, a scalar.
}
\description{
Entropy of a Gamma distribution
}
