% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_and_KL.r
\name{entropy_pool}
\alias{entropy_pool}
\title{Compute the entropy of a general pool of distributions}
\usage{
entropy_pool(D, alpha)
}
\arguments{
\item{D}{list containing the density functions.}

\item{alpha}{vector of weights, which must lie in a simplex (can be a scalar).}
}
\value{
the entropy, a scalar.
}
\description{
Compute the entropy of a general pool of distributions
}
