% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_and_KL.r
\name{entropy_surface_beta}
\alias{entropy_surface_beta}
\title{Compute the entropy of Beta distribution on a grid of shape parameters.}
\usage{
entropy_surface_beta(av, bv, N = 100)
}
\arguments{
\item{av}{a vector of shape parameters, all greater than zero.}

\item{bv}{a vector of shape parameters, all greater than zero.}

\item{N}{Number of grid points to compute.}
}
\value{
a list containing a matrix with the entropies computed on a N x N grid, as
well as vectors containing values of \code{a} and \code{b} in the range of \code{av} and \code{bv}.
}
\description{
Compute the entropy of Beta distribution on a grid of shape parameters.
}
