% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_likelihoods.r
\name{get_normal_weights}
\alias{get_normal_weights}
\title{Compute BMA weights for the normal model}
\usage{
get_normal_weights(obs, sigma, ms, vs)
}
\arguments{
\item{obs}{vector of observations.}

\item{sigma}{fixed sampling standard deviation.}

\item{ms}{prior means.}

\item{vs}{prior variances.}
}
\value{
a vector of weights of the same size as \code{ms} and \code{vs}.
}
\description{
Compute BMA weights for the normal model
}
