% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_and_KL.r
\name{kl_beta}
\alias{kl_beta}
\title{Kullback-Leibler divergence a Beta and the pool.}
\usage{
kl_beta(astar, bstar, ai, bi, type = c("pf", "fp"))
}
\arguments{
\item{astar}{shape parameter of the pool \code{pi}.}

\item{bstar}{shape parameter of the pool \code{pi}.}

\item{ai}{shape parameter of \code{f}.}

\item{bi}{shape parameter of \code{f}.}

\item{type}{if \code{type = pf}, computes KL(pi||f), whereas
\code{type = fp}, computes KL(f || pi).}
}
\value{
the KL divergence. A scalar.
}
\description{
Kullback-Leibler divergence a Beta and the pool.
}
\details{
Here pi ~ Beta(astar, bstar) and f ~ Beta(ai, bi).
}
