% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_and_KL.r
\name{kl_gamma}
\alias{kl_gamma}
\title{Kullback-Leibler divergence a Gamma and the pool.}
\usage{
kl_gamma(astar, bstar, ai, bi, type = c("pf", "fp"))
}
\arguments{
\item{astar}{shape parameter of the pool \code{pi}.}

\item{bstar}{rate parameter of the pool \code{pi}.}

\item{ai}{shape parameter of \code{f}.}

\item{bi}{rate parameter of \code{f}.}

\item{type}{if \code{type = pf}, computes KL(pi||f), whereas
\code{type = fp}, computes KL(f || pi).}
}
\value{
the KL divergence. A scalar.
}
\description{
Kullback-Leibler divergence a Gamma and the pool.
}
\details{
Here pi ~ Gamma(astar, bstar) and f ~ Gamma(ai, bi)
}
