% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{mean_ci}
\alias{mean_ci}
\title{Return mean and 100*alpha\% quantiles.}
\usage{
mean_ci(x, alpha = 0.95, na.rm = TRUE)
}
\arguments{
\item{x}{vector of values to be summarised.}

\item{alpha}{probability level. A number between 0 and 1.}

\item{na.rm}{logical.  If \code{TRUE} \code{NA}s are excluded.}
}
\value{
a data.frame with mean and upper and lower quantiles.
}
\description{
Return mean and 100*alpha\% quantiles.
}
