% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_likelihoods.r
\name{ml_beta}
\alias{ml_beta}
\title{Marginal likelihood under a Beta-binomial model}
\usage{
ml_beta(yi, ni, a, b, log = FALSE)
}
\arguments{
\item{yi}{number of successes.}

\item{ni}{number of trials.}

\item{a}{prior shape1 parameter.}

\item{b}{prior shape2 parameter.}

\item{log}{logical. Whether to return the log marginal likelihood.}
}
\value{
the marginal likelihood.
}
\description{
Marginal likelihood under a Beta-binomial model
}
