% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_likelihoods.r
\name{normal_mean_marg_like}
\alias{normal_mean_marg_like}
\title{Marginal likelihood for a Normal-Normal model}
\usage{
normal_mean_marg_like(y, sigma, m, v, log = TRUE)
}
\arguments{
\item{y}{a vector of observations.}

\item{sigma}{fixed sampling standard deviation.}

\item{m}{prior mean.}

\item{v}{prior variance.}

\item{log}{logical. Whether to return the log marginal likelihood.}
}
\value{
the marginal likelihood.
}
\description{
Marginal likelihood for a Normal-Normal model
}
