% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_and_KL.r
\name{optentbeta}
\alias{optentbeta}
\title{Loss function for maximising the entropy of a Beta pool.}
\usage{
optentbeta(alpha, ap, bp)
}
\arguments{
\item{alpha}{vector of weights, which must lie in a simplex (can be a scalar).}

\item{ap}{vector of shape parameters, all larger than zero.}

\item{bp}{vector of shape parameters, all larger than zero.}
}
\value{
the entropy of the pool, a scalar.
}
\description{
Loss function for maximising the entropy of a Beta pool.
}
