% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_and_KL.r
\name{optentbeta_inv}
\alias{optentbeta_inv}
\title{Loss function for maximising the entropy of a Beta pool.}
\usage{
optentbeta_inv(alpha.inv, ap, bp)
}
\arguments{
\item{alpha.inv}{vector of weights mapped onto R^(K-1).}

\item{ap}{vector of shape parameters, all larger than zero.}

\item{bp}{vector of shape parameters, all larger than zero.}
}
\value{
the negative of the entropy, a scalar.
}
\description{
Loss function for maximising the entropy of a Beta pool.
}
