% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_and_KL.r
\name{optentgauss}
\alias{optentgauss}
\title{Loss function for maximising the entropy of a Gaussian pool.}
\usage{
optentgauss(alpha, mp, vp)
}
\arguments{
\item{alpha}{vector of weights, which must lie in a simplex (can be a scalar).}

\item{mp}{vector of means.}

\item{vp}{vector of variances, all larger than zero.}
}
\value{
the entropy of the pool, a scalar.
}
\description{
Loss function for maximising the entropy of a Gaussian pool.
}
