% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_and_KL.r
\name{optklbeta}
\alias{optklbeta}
\title{Loss function for minimising the sum of KL divergences of a Beta pool}
\usage{
optklbeta(alpha, ap, bp, type)
}
\arguments{
\item{alpha}{vector of weights, which must lie in a simplex (can be a scalar).}

\item{ap}{vector of shape parameters, all larger than zero.}

\item{bp}{vector of shape parameters, all larger than zero.}

\item{type}{if \code{type = pf}, computes KL(pi||f), whereas
\code{type = fp}, computes KL(f || pi). See \code{\link[logPooR]{kl_beta}}.}
}
\value{
a vector containing the KL divergences.
}
\description{
Loss function for minimising the sum of KL divergences of a Beta pool
}
