% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_and_KL.r
\name{optklbeta_inv}
\alias{optklbeta_inv}
\title{Loss function for maximising the entropy of a Beta pool.}
\usage{
optklbeta_inv(alpha.inv, ap, bp, type = "pf")
}
\arguments{
\item{alpha.inv}{vector of weights mapped onto R^(K-1).}

\item{ap}{vector of shape parameters, all larger than zero.}

\item{bp}{vector of shape parameters, all larger than zero.}

\item{type}{if \code{type = pf}, computes KL(pi||f), whereas
\code{type = fp}, computes KL(f || pi). See \code{\link[logPooR]{kl_beta}}.}
}
\value{
the sum of KL divergences, a scalar.
}
\description{
Loss function for maximising the entropy of a Beta pool.
}
