% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_and_KL.r
\name{optklgamma_inv}
\alias{optklgamma_inv}
\title{Loss function for minimising the sum of KL divergences of a Gamma pool.}
\usage{
optklgamma_inv(alpha.inv, ap, bp, type = "pf")
}
\arguments{
\item{alpha.inv}{vector of weights mapped onto R^(K-1).}

\item{ap}{vector of shape parameters, all larger than zero.}

\item{bp}{vector of rate parameters, all larger than zero.}

\item{type}{if \code{type = pf}, computes KL(pi||f), whereas
\code{type = fp}, computes KL(f || pi). See  \code{\link[logPooR]{kl_gamma}}.}
}
\value{
the sum of KL divergences, a scalar.
}
\description{
Loss function for minimising the sum of KL divergences of a Gamma pool.
}
