% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_and_KL.r
\name{optklgauss}
\alias{optklgauss}
\title{Loss function for minimising the sum of KL divergences of a Gaussian pool.}
\usage{
optklgauss(alpha, mp, vp, type)
}
\arguments{
\item{alpha}{vector of weights, which must lie in a simplex (can be a scalar).}

\item{mp}{vector of means.}

\item{vp}{vector of variances, all larger than zero.}

\item{type}{if \code{type = pf}, computes KL(pi||f), whereas
\code{type = fp}, computes KL(f || pi). See  \code{\link[logPooR]{kl_gauss}}.}
}
\value{
a vector containing the KL divergences.
}
\description{
Loss function for minimising the sum of KL divergences of a Gaussian pool.
}
