% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_and_KL.r
\name{optklgauss_inv}
\alias{optklgauss_inv}
\title{Loss function for minimising the sum of KL divergences of a Gaussian pool.}
\usage{
optklgauss_inv(alpha.inv, mp, vp, type = "pf")
}
\arguments{
\item{alpha.inv}{vector of weights mapped onto R^(K-1).}

\item{mp}{vector of means.}

\item{vp}{vector of variances, all larger than zero.}

\item{type}{if \code{type = pf}, computes KL(pi||f), whereas
\code{type = fp}, computes KL(f || pi). See  \code{\link[logPooR]{kl_gauss}}.}
}
\value{
a vector containing the KL divergences.
}
\description{
Loss function for minimising the sum of KL divergences of a Gaussian pool.
}
