% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.r
\name{pool_par}
\alias{pool_par}
\title{Obtain the parameters of the pooled distribution (Beta and Gamma).}
\usage{
pool_par(alpha, a, b)
}
\arguments{
\item{alpha}{vector of weights, which must lie in a simplex (can be a scalar).}

\item{a}{vector of 'a' parameters, same size as 'alpha' and 'b'.}

\item{b}{vector of 'b' parameters.}
}
\value{
a vector with two entries, a_star and b_star.
}
\description{
Obtain the parameters of the pooled distribution (Beta and Gamma).
}
