% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.r
\name{pool_par_gauss}
\alias{pool_par_gauss}
\title{Obtain the parameters of the pooled distribution (Gaussian).}
\usage{
pool_par_gauss(alpha, m, v)
}
\arguments{
\item{alpha}{vector of weights, which must lie in a simplex (can be a scalar).}

\item{m}{vector of means, same dimension as 'alpha' and 'v'.}

\item{v}{vector of \emph{variances}.}
}
\value{
pooled mean and \emph{STANDARD DEVIATION}.
}
\description{
Obtain the parameters of the pooled distribution (Gaussian).
}
