% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{rlogisticnorm}
\alias{rlogisticnorm}
\title{Generate samples from a multivariate logistic normal.}
\usage{
rlogisticnorm(N, m, Sigma)
}
\arguments{
\item{N}{number of draws.}

\item{m}{is a vector of means. If a scalar is given, it is recycled \code{K} times.}

\item{Sigma}{the variance-covariance matrix.}
}
\description{
Generate samples from a multivariate logistic normal.
}
