% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.r
\name{tpool_positive}
\alias{tpool_positive}
\title{Normalising constant of the pooled distributions with support on (0, Inf).}
\usage{
tpool_positive(alpha, D, trapez = FALSE, lwr_num = 0, upr_num = 1e+05)
}
\arguments{
\item{alpha}{vector of weights, which must lie in a simplex (can be a scalar).}

\item{D}{list containing the density functions.}

\item{trapez}{logical. Whether to use the trapezoid method of integration.}

\item{lwr_num}{approximate lower limit of integration if 'trapezoid' is \code{TRUE}.}

\item{upr_num}{approximate upper limit of integration if 'trapezoid' is \code{TRUE}.}
}
\value{
numeric. The normalising constant.
}
\description{
Normalising constant of the pooled distributions with support on (0, Inf).
}
