% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.r
\name{tpool_unit}
\alias{tpool_unit}
\title{Normalising constant of the pooled distributions with support on (0, 1).}
\usage{
tpool_unit(alpha, D, lwr_num = 1e-10, upr_num = 1 - 1e-10)
}
\arguments{
\item{alpha}{vector of weights, which must lie in a simplex (can be a scalar).}

\item{D}{list containing the density functions.}

\item{lwr_num}{approximate lower limit of integration if 'trapezoid' is \code{TRUE}.}

\item{upr_num}{approximate upper limit of integration if 'trapezoid' is \code{TRUE}.}
}
\value{
numeric. The normalising constant.
}
\description{
Normalising constant of the pooled distributions with support on (0, 1).
}
